+++ source/scripts/serviceWorker/sw.js
@@ -1,0 +1,5 @@
+/* eslint-env serviceworker */
+
+import initCache from './cache'
+
+initCache()